// fe/src/lib/api/generated/model/cSMCC3Value.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type CSMCC3Value = typeof CSMCC3Value[keyof typeof CSMCC3Value];


export const CSMCC3Value = {
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
} as const;