// fe/src/lib/api/generated/model/cSMClientTypeValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type CSMClientTypeValue = typeof CSMClientTypeValue[keyof typeof CSMClientTypeValue];


export const CSMClientTypeValue = {
  CITIZEN: 'CITIZEN',
  BUSINESS: 'BUSINESS',
  GOVERNMENT: 'GOVERNMENT',
} as const;