// fe/src/lib/api/generated/model/customerFeedbackAccommodatedByValue.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * COMPASS API
 * The version-one backend API for COMPASS.
 * OpenAPI spec version: 1.0.0
 */

export type CustomerFeedbackAccommodatedByValue = typeof CustomerFeedbackAccommodatedByValue[keyof typeof CustomerFeedbackAccommodatedByValue];


export const CustomerFeedbackAccommodatedByValue = {
  STUDENT_ASSISTANT: 'STUDENT_ASSISTANT',
  CLERK_PERSONNEL: 'CLERK_PERSONNEL',
} as const;